cchar form 3 done
