+++ components/create-character/CChar-Form-3.vue
@@ -1,8 +1,9 @@
 // Appearance
 
 <script setup lang="ts">
-import type { FormError, FormSubmitEvent } from "#ui/types";
+import type { FormSubmitEvent } from "#ui/types";
 import type { CharacterForm } from "~/types/types";
+import { characterValidate3 } from "~/utils/character-validation-3";
 import { useAuthStore, useFormStore } from "#imports";
 
 const config = useRuntimeConfig();
@@ -24,25 +25,6 @@ const state = reactive({
   disabled: false,
   loading: false,
 });
-
-const validate = (state: CharacterForm): FormError[] => {
-  const errors = [];
-  if (state.hair && state.hair.length > 25)
-    errors.push({ path: "hair", message: "Must be less than 25 characters" });
-  if (state.skin && state.skin.length > 25)
-    errors.push({ path: "skin", message: "Must be less than 25 characters" });
-  if (state.eyes && state.eyes.length > 25)
-    errors.push({ path: "eyes", message: "Must be less than 25 characters" });
-  if (state.height && state.height < 1)
-    errors.push({ path: "height", message: "Must be at least 12 inches" });
-  if (state.gender && state.gender.length > 25)
-    errors.push({ path: "gender", message: "Must be less than 25 characters" });
-  if (state.age && state.age < 1)
-    errors.push({ path: "age", message: "Must be at least 1 year old" });
-  if (state.weight && state.weight < 1)
-    errors.push({ path: "weight", message: "Must be at least 1lb" });
-  return errors;
-};
 
 function updateFormStore(fields: CharacterForm) {
   Object.entries(fields).forEach(([field, value]) => {
@@ -113,7 +95,11 @@ async function onSubmit(event: FormSubmitEvent<CharacterForm>) {
 </script>
 
 <template>
-  <UForm :validate="validate" :state="state" @submit.prevent="onSubmit">
+  <UForm
+    :validate="characterValidate3"
+    :state="state"
+    @submit.prevent="onSubmit"
+  >
     <div class="grid grid-cols-2 gap-4">
       <UFormGroup label="Hair" name="hair">
         <UInput v-model="state.hair" :disabled="state.disabled" type="text" />
